+++ scripts/ci/pylint_todo.txt
@@ -50,7 +50,7 @@
 ./airflow/contrib/operators/oracle_to_oracle_transfer.py
 ./airflow/contrib/operators/qubole_check_operator.py
 ./airflow/contrib/operators/redis_publish_operator.py
-./airflow/contrib/operators/s3_copy_object_operator.py
+./airflow/providers/amazon/aws/operators/s3_copy_object.py
 ./airflow/contrib/operators/s3_delete_objects_operator.py
 ./airflow/contrib/operators/s3_list_operator.py
 ./airflow/contrib/operators/s3_to_gcs_operator.py
