+++ tests/test_core_to_contrib.py
@@ -938,6 +938,10 @@ OPERATOR = [
         'airflow.providers.amazon.aws.operators.emr_terminate_job_flow.EmrTerminateJobFlowOperator',
         'airflow.contrib.operators.emr_terminate_job_flow_operator.EmrTerminateJobFlowOperator',
     ),
+    (
+        'airflow.providers.amazon.aws.operators.s3_copy_object.S3CopyObjectOperator',
+        'airflow.contrib.operators.s3_copy_object_operator.S3CopyObjectOperator',
+    ),
 ]
 
 SENSOR = [
